143 fix fix site detail part

+++ ServerMonitor/Services/RequestServices/BasicRequest.cs
@@ -2,17 +2,20 @@
 
 namespace ServerMonitor.Services.RequestServices
 {
+    /// <summary>
+    /// 请求模块的顶层类 创建者:xb 创建时间: 2018/04
+    /// </summary>
     public abstract class BasicRequest
     {
         public readonly double ErrorQuality = 1.5;
         /// <summary>
-        /// 请求创建的时间
+        /// 请求创建的时间 
         /// </summary>
         private DateTime createTime;
         /// <summary>
         /// 请求花费的时间
         /// </summary>
-        private short timeCost;
+        private int timeCost;
         /// <summary>
         /// 定义的请求超时时长
         /// </summary>
@@ -31,7 +34,7 @@ namespace ServerMonitor.Services.RequestServices
         private Exception errorException;       
 
         public DateTime CreateTime { get => createTime; set => createTime = value; }
-        public short TimeCost { get => timeCost; set => timeCost = value; }
+        public int TimeCost { get => timeCost; set => timeCost = value; }
         public short OverTime { get => overTime; set => overTime = value; }
         public string Status { get => status; set => status = value; }
         public string Others { get => others; set => others = value; }
