+++ ServerMonitor/Services/RequestServices/IcmpRequest.cs
@@ -62,7 +62,7 @@ namespace ServerMonitor.Services.RequestServices
                         //Color = "0";
                         request.Color = "0";
                         request.Status = "1001";
-                        request.TimeCost = (short)(request.OverTime * request.ErrorQuality);
+                        request.TimeCost = (int)(request.OverTime * request.ErrorQuality);
                         Requests.Add(request);
                         return false;
                     }
@@ -88,7 +88,7 @@ namespace ServerMonitor.Services.RequestServices
                         //Color = "0";
                         request.Color = "0";
                         request.Status = "1001";
-                        request.TimeCost = (short)(request.OverTime * request.ErrorQuality);
+                        request.TimeCost = (int)(request.OverTime * request.ErrorQuality);
                         //string backJson = JsonConvert.SerializeObject(information);
                         //backData.Add("1", backJson);
                         Requests.Add(request);
@@ -115,8 +115,8 @@ namespace ServerMonitor.Services.RequestServices
                             //Color = "0";//访问被拒绝
                             request.Color = "0";
                             request.Status = "1001";
-                            request.TimeCost = (short)(request.OverTime * request.ErrorQuality);
-                            //TimeCost = (short)stopwatch.ElapsedMilliseconds;
+                            request.TimeCost = (int)(request.OverTime * request.ErrorQuality);
+                            //TimeCost = stopwatch.ElapsedMilliseconds;
                             Requests.Add(request);
                             return false;
                         }
@@ -151,7 +151,7 @@ namespace ServerMonitor.Services.RequestServices
                                 //string backJson = JsonConvert.SerializeObject(information);
                                 //backData.Add(i.ToString(), backJson);
                                 //return backData;
-                                request.TimeCost = (short)(request.OverTime * request.ErrorQuality);
+                                request.TimeCost = (int)(request.OverTime * request.ErrorQuality);
                                 Requests.Add(request);
                                 break;
                             }
@@ -168,7 +168,7 @@ namespace ServerMonitor.Services.RequestServices
                                 //string backJson = JsonConvert.SerializeObject(information);
                                 //backData.Add(i.ToString(), backJson);
                                 //return backData;
-                                request.TimeCost = (short)(request.OverTime * request.ErrorQuality);
+                                request.TimeCost = (int)(request.OverTime * request.ErrorQuality);
                                 Requests.Add(request);
                                 break;
                             }
@@ -188,7 +188,7 @@ namespace ServerMonitor.Services.RequestServices
                                 //backData.Add(i.ToString(), backJson);
                                 request.Color = "1";
                                 request.Status = "1000";
-                                request.TimeCost = (short)stopwatch.ElapsedMilliseconds;
+                                request.TimeCost = (int)stopwatch.ElapsedMilliseconds;
                                 Requests.Add(request);
                                 break;
                             }
@@ -206,7 +206,7 @@ namespace ServerMonitor.Services.RequestServices
                                 //string backJson = JsonConvert.SerializeObject(information);
                                 //backData.Add(i.ToString(), backJson);
                                 //return backData;
-                                request.TimeCost = (short)stopwatch.ElapsedMilliseconds;
+                                request.TimeCost = (int)stopwatch.ElapsedMilliseconds;
                                 Requests.Add(request);
                                 break;
                             }
@@ -224,7 +224,7 @@ namespace ServerMonitor.Services.RequestServices
                         request.Status = "1001";
                         //string backJson = JsonConvert.SerializeObject(information);
                         //backData.Add(i.ToString(), backJson);
-                        request.TimeCost = (short)(request.OverTime * request.ErrorQuality);
+                        request.TimeCost = (int)(request.OverTime * request.ErrorQuality);
                         request.ErrorException = ex;
                         Requests.Add(request);
                         return false;
@@ -245,7 +245,7 @@ namespace ServerMonitor.Services.RequestServices
                 };//创建一个请求对象
                 request.Color = "0";
                 request.Status = "1001";
-                request.TimeCost = (short)(request.OverTime * request.ErrorQuality);       
+                request.TimeCost = (int)(request.OverTime * request.ErrorQuality);       
                 return false;
             }
         }
